# swagger/jmarkets.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "JMarkets API Documentation",
    "version": "1.0.0",
    "description": "Multi-tenant marketplace platform API with CMS, authentication, RBAC, and auto-deployment",
    "contact": {
      "name": "JMarkets Support",
      "url": "https://j-markets.jcampos.dev"
    }
  },
  "servers": [
    {
      "url": "https://markets-api.jcampos.dev",
      "description": "Production API"
    },
    {
      "url": "http://localhost:5000",
      "description": "Development API"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
...
  },
  "paths": {
...
    "/api/organizations/{id}/settings": {
      "put": {
        "summary": "Update organization settings",
        "tags": [
          "Organizations"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Settings updated"
          },
          "404": {
            "description": "Organization not found"
          }
        }
      }
    },
...
    "/api/users/{userId}/organization/{orgId}/settings/payment": {
      "get": {
        "summary": "Get payment settings for an organization",
        "tags": [
          "Settings"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "orgId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment settings"
          },
          "404": {
            "description": "Payment settings not found"
          }
        }
      },
      "put": {
        "summary": "Update payment settings for an organization",
        "tags": [
          "Settings"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "orgId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "currency": {
                    "type": "string"
                  },
                  "stripeEnabled": {
                    "type": "boolean"
                  },
                  "stripePublishableKey": {
                    "type": "string"
                  },
                  "stripeSecretKey": {
                    "type": "string"
                  },
                  "cashOnDeliveryEnabled": {
                    "type": "boolean"
                  },
                  "bankTransferEnabled": {
                    "type": "boolean"
                  },
                  "bankAccountDetails": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment settings updated successfully"
          },
          "404": {
            "description": "Payment settings not found"
          }
        }
      }
    },
...
  },
...
}